requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.30.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.1.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.1
questionary==2.1.0
requests==2.32.3
simpleeval==1.0.3
SQLAlchemy==2.0.36
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
